wip: port the diff command to the compare API
