+++ src/cli/src/cmd_setup.rs
@@ -815,13 +815,13 @@ pub fn compare() -> Command {
             .help("Second resource, in format `file` or `file:revision`")
             .index(2))
         .arg(Arg::new("keys")
-            .required(true)
+            .required(false)
             .long("keys")
             .help("Comma-separated list of columns to compare on. If not specified, all columns are used for comparison.")
             .use_value_delimiter(true)
             .action(clap::ArgAction::Set))
         .arg(Arg::new("targets")
-            .required(true)
+            .required(false)
             .long("targets")
             .help("Comma-separated list of columns in which to view changes. If not specified, all columns are viewed")
             .use_value_delimiter(true)
